organized file structure to include models/ and services/
