+++ src/app/auditions/auditions.component.css
@@ -10,7 +10,8 @@
 .sidebar {
   position: sticky;
   top: 12%;
-  width: 24%;
+  /* width: 24%; */
+  padding: 30px;
   align-self: flex-start; /* Aligns the sidebar to the top of the container */
 }
