+++ src/app/services/members.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { Member } from './member';
+import { Member } from '../models/member';
 
 
 @Injectable({
