+++ src/app/services/photos.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { Photo } from './photos';
+import { Photo } from '../models/photos';
 
 @Injectable({
   providedIn: 'root',
