+++ src/app/theme-toggle/theme-toggle.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { ThemeService } from '../theme.service';
+import { ThemeService } from '../services/theme.service';
 
 @Component({
   selector: 'app-theme-toggle',
